Create IO_MODULE
I now have an IO module to handle those few variables and all
associated functions and error routines. This was a pretty big
change beacause I removed CVAR and CFLD from the APM_MODULE
as well as shifting unit numbers over to the module and moving
READ_MASTER to the main program file.