Move all variables defined in MAP_MODULE back to APM_MODULE
I had good reasoning to initially move those "aperture map" variables
to the map module but I would much rather have a clean, independent
module instead of one cluttered with other data.

The apm_module is the place for "model-specific" variables to be
collected. The IO_MODULE is so far a standing exception and is not
generalized. However, the io_module has no other purpose than to
serve the modeling package as a whole. Whereas the map module could
be used by other routines (i.e. Python package)