Remove usage of IVTK in WRITE_VTK_FILE
This seems counter intuitive but the more I can parameterize a
subroutine the better since it reduces how much I have to rely
on a given dependency. It also opens up the door for me to output
two VTK files if I'd ever want to. Say one in SI units and the other
in the desired output units